backend/requirements.txt
# backend
fastapi
uvicorn

# webscraper 
python-dotenv
beautifulsoup4
readability-lxml
requests
trafilatura
newspaper3k
finnhub-python
lxml
feedparser
supabase

#forecasting
prophet
yfinance
pandas
statsmodels
matplotlib
xgboost
scikit-learn

# PyTorch - install with CUDA support separately if needed:
# pip install torch==2.0.1+cu118 --index-url https://download.pytorch.org/whl/cu118
torch

# Add Transformers library
transformers